:sparkles: feat: add janus color scheme
