+++ frontend/src/theme/index.ts
@@ -1,5 +1,8 @@
 import { CssVarsThemeOptions } from "@mui/material";
 
+// augments the Theme interface with CSS variables and color schemes
+import type {} from "@mui/material/themeCssVarsAugmentation";
+
 import { colorSchemes } from "./colorSchemes";
 import { components } from "./components";
 import { shadows } from "./shadows";
